pyproject.toml
[tool.poetry]
name = "shadow-nexus"
version = "0.1.0"
description = "Advanced AI Command & Control System with Forex Trading, Data Retrieval, and Information Gathering capabilities"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "AthenaMist" },
    { include = "Phantom" },
    { include = "SovereignCore" },
    { include = "Command_Network" },
    { include = "Core_Control" },
    { include = "Security" },
    { include = "IoT_Control" },
    { include = "Config" }
]

[tool.poetry.dependencies]
python = "^3.11"
# AWS
boto3 = "^1.34.0"
aws-amplify = "^1.0.0"

# Discord
"discord.py" = "^2.3.2"

# Telegram
python-telegram-bot = "^20.7"

# Email
aiosmtplib = "^3.0.1"

# Data Processing
pandas = "^2.1.4"
numpy = "^1.26.2"
beautifulsoup4 = "^4.12.2"
feedparser = "^6.0.10"
pydantic = "^2.5.2"

# HTTP & Async
aiohttp = "^3.9.1"
requests = "^2.31.0"
fake-useragent = "^1.4.0"

# Encryption
pycryptodome = "^3.19.0"

# Testing
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
moto = "^4.2.11"

# Logging & Monitoring
structlog = "^23.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
isort = "^5.13.2"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=. --cov-report=term-missing" 